Day 22 optimization
Cut out tuples to save a large amount of time

This is also more correct/accurate in that it stores the total for all sequences across all buyers rather than requiring the best sequence to be in the first buyer's set.

Finally, it turns out that the float divide/truncate was unnecessary and also costing some time.

In the debugger, part 2 now finishes in around 675ms. Debug config outside the debugger is done in 360ms, and Release in 200ms. Good enough for now, I think.